+++ docs/lib/Components/TooltipsPage.js
@@ -4,6 +4,8 @@ import { PrismCode } from 'react-prism';
 import Helmet from 'react-helmet';
 import TooltipExample from '../examples/Tooltip';
 const TooltipExampleSource = require('!!raw!../examples/Tooltip');
+import TooltipAutoHideExample from '../examples/TooltipAutoHide';
+const TooltipExampleAutoHideSource = require('!!raw!../examples/TooltipAutoHide');
 import TooltipExampleMulti from '../examples/TooltipMulti';
 const TooltipExampleMultiSource = require('!!raw!../examples/TooltipMulti');
 
@@ -34,6 +36,13 @@ export default class TooltipsPage extends React.Component {
   // target div ID, popover is attached to this element
   tether: PropTypes.oneOfType([PropTypes.object, PropTypes.bool]),
   // optionally overide tether config http://tether.io/#options
+  delay: PropTypes.oneOfType([
+    PropTypes.shape({ show: PropTypes.number, hide: PropTypes.number }),
+    PropTypes.number
+  ]),
+  // optionally override show/hide delays - default { show: 0, hide: 250 }
+  autohide: PropTypes.bool,
+  // optionally hide tooltip when hovering over tooltip content - default true
   placement: PropTypes.oneOf([
     'top',
     'bottom',
@@ -55,6 +64,15 @@ export default class TooltipsPage extends React.Component {
   // convenience attachments for popover
   // examples http://github.hubspot.com/tooltip/docs/welcome/
 }`}
+          </PrismCode>
+        </pre>
+        <h3>Tooltip Disable Autohide</h3>
+        <div className="docs-example">
+          <TooltipAutoHideExample />
+        </div>
+        <pre>
+          <PrismCode className="language-jsx">
+            {TooltipExampleAutoHideSource}
           </PrismCode>
         </pre>
         <h3>Tooltips List</h3>