+++ docs/lib/Components/index.js
@@ -11,7 +11,9 @@ const ComponentLink = (props) => {
     </NavItem>
   );
 };
-
+const propTypes = {
+  children: React.PropTypes.node
+};
 
 class Components extends React.Component {
   constructor(props) {
@@ -95,6 +97,14 @@ class Components extends React.Component {
           name: 'Pagination',
           to: '/components/pagination/'
         },
+        {
+          name: 'Tabs',
+          to: '/components/tabs/'
+        },
+        {
+          name: 'Jumbotron',
+          to: '/components/jumbotron/'
+        }
       ]
     };
   }
@@ -123,5 +133,5 @@ class Components extends React.Component {
     );
   }
 }
-
+Components.propTypes = propTypes;
 export default Components;