+++ docs/lib/examples/Tooltip.js
@@ -1,6 +1,6 @@
 /* eslint react/no-multi-comp: 0, react/prop-types: 0 */
 import React from 'react';
-import { Button, Tooltip } from 'reactstrap';
+import { Tooltip } from 'reactstrap';
 
 export default class Example extends React.Component {
   constructor(props) {