+++ docs/lib/routes.js
@@ -20,6 +20,8 @@ import ModalsPage from './Components/ModalsPage';
 import CardPage from './Components/CardPage';
 import TablesPage from './Components/TablesPage';
 import PaginationPage from './Components/PaginationPage';
+import TabsPage from './Components/TabsPage';
+import JumbotronPage from './Components/JumbotronPage';
 import NotFound from './NotFound';
 import Components from './Components';
 import UI from './UI';
@@ -48,6 +50,8 @@ const routes = (
       <Route path="navbar/" component={NavbarPage} />
       <Route path="media/" component={MediaPage} />
       <Route path="pagination/" component={PaginationPage} />
+      <Route path="tabs/" component={TabsPage} />
+      <Route path="jumbotron/" component={JumbotronPage} />
     </Route>
     <Route path="*" component={NotFound} />
   </Route>