+++ src/index.js
@@ -58,6 +58,9 @@ import Media from './Media';
 import Pagination from './Pagination';
 import PaginationItem from './PaginationItem';
 import PaginationLink from './PaginationLink';
+import TabContent from './TabContent';
+import TabPane from './TabPane';
+import Jumbotron from './Jumbotron';
 
 export {
   Container,
@@ -120,4 +123,7 @@ export {
   Pagination,
   PaginationItem,
   PaginationLink,
+  TabContent,
+  TabPane,
+  Jumbotron
 };